+++ dart/test/sentry_test.dart
@@ -5,10 +5,9 @@ import 'package:sentry/src/dart_exception_type_identifier.dart';
 import 'package:sentry/src/event_processor/deduplication_event_processor.dart';
 import 'package:test/test.dart';
 
-import 'fake_platform_checker.dart';
 import 'mocks.dart';
 import 'mocks/mock_integration.dart';
-import 'mocks/mock_platform_checker.dart';
+import 'mocks/mock_runtime_checker.dart';
 import 'mocks/mock_sentry_client.dart';
 import 'test_utils.dart';
 
@@ -372,7 +371,7 @@ void main() {
 
     test('should set options.debug to true when in debug mode', () async {
       final options = defaultTestOptions();
-      options.platformChecker = MockPlatformChecker(isDebug: true);
+      options.runtimeChecker = MockRuntimeChecker(isDebug: true);
 
       expect(options.debug, isFalse);
       await Sentry.init(
@@ -386,7 +385,7 @@ void main() {
 
     test('should respect user options.debug when in debug mode', () async {
       final options = defaultTestOptions();
-      options.platformChecker = MockPlatformChecker(isDebug: true);
+      options.runtimeChecker = MockRuntimeChecker(isDebug: true);
 
       expect(options.debug, isFalse);
       await Sentry.init(
@@ -402,7 +401,7 @@ void main() {
     test('should leave options.debug unchanged when not in debug mode',
         () async {
       final options = defaultTestOptions();
-      options.platformChecker = MockPlatformChecker(isDebug: false);
+      options.runtimeChecker = MockRuntimeChecker(isDebug: false);
 
       expect(options.debug, isFalse);
       await Sentry.init(
@@ -443,7 +442,8 @@ void main() {
   });
 
   test('options.environment debug', () async {
-    final sentryOptions = defaultTestOptions(FakePlatformChecker.debugMode());
+    final sentryOptions =
+        defaultTestOptions(checker: MockRuntimeChecker(isDebug: true));
     await Sentry.init(
       (options) {
         options.dsn = fakeDsn;
@@ -455,7 +455,8 @@ void main() {
   });
 
   test('options.environment profile', () async {
-    final sentryOptions = defaultTestOptions(FakePlatformChecker.profileMode());
+    final sentryOptions =
+        defaultTestOptions(checker: MockRuntimeChecker(isProfile: true));
 
     await Sentry.init(
       (options) {
@@ -468,7 +469,8 @@ void main() {
   });
 
   test('options.environment production (defaultEnvironment)', () async {
-    final sentryOptions = defaultTestOptions(FakePlatformChecker.releaseMode());
+    final sentryOptions =
+        defaultTestOptions(checker: MockRuntimeChecker(isRelease: true));
     await Sentry.init(
       (options) {
         options.dsn = fakeDsn;
@@ -480,7 +482,8 @@ void main() {
   });
 
   test('options.logger is set by setting the debug flag', () async {
-    final sentryOptions = defaultTestOptions(FakePlatformChecker.debugMode());
+    final sentryOptions =
+        defaultTestOptions(checker: MockRuntimeChecker(isDebug: true));
 
     await Sentry.init(
       (options) {
@@ -505,9 +508,9 @@ void main() {
     });
 
     test('throw is handled and logged', () async {
-      // Use release mode in platform checker to avoid additional log
+      // Use release mode in runtime checker to avoid additional log
       final sentryOptions =
-          defaultTestOptions(FakePlatformChecker.releaseMode())
+          defaultTestOptions(checker: MockRuntimeChecker(isRelease: true))
             ..automatedTestMode = false
             ..debug = true
             ..logger = fixture.mockLogger;